+++ assets.cmake
@@ -178,6 +178,7 @@ qt_add_resources(green "images"
         assets/svg2/gear.svg
         assets/svg2/hand.svg
         assets/svg2/house.svg
+        assets/svg2/import.svg
         assets/svg2/incoming.svg
         assets/svg2/info.svg
         assets/svg2/info_red.svg